Move FilterLogs function to a new package so that mel code in nitro is wasm compatible
