+++ eth/filters/filter.go
@@ -19,12 +19,11 @@ package filters
 import (
 	"context"
 	"errors"
-	"math/big"
-	"slices"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/bloombits"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/melfilters"
 	"github.com/ethereum/go-ethereum/rpc"
 )
 
@@ -291,7 +290,7 @@ func (f *Filter) checkMatches(ctx context.Context, header *types.Header) ([]*typ
 	if err != nil {
 		return nil, err
 	}
-	logs := FilterLogs(cached.logs, nil, nil, f.addresses, f.topics)
+	logs := melfilters.FilterLogs(cached.logs, nil, nil, f.addresses, f.topics)
 	if len(logs) == 0 {
 		return nil, nil
 	}
@@ -311,41 +310,6 @@ func (f *Filter) checkMatches(ctx context.Context, header *types.Header) ([]*typ
 		logs[i] = &logcopy
 	}
 	return logs, nil
-}
-
-// FilterLogs creates a slice of logs matching the given criteria.
-func FilterLogs(logs []*types.Log, fromBlock, toBlock *big.Int, addresses []common.Address, topics [][]common.Hash) []*types.Log {
-	var check = func(log *types.Log) bool {
-		if fromBlock != nil && fromBlock.Int64() >= 0 && fromBlock.Uint64() > log.BlockNumber {
-			return false
-		}
-		if toBlock != nil && toBlock.Int64() >= 0 && toBlock.Uint64() < log.BlockNumber {
-			return false
-		}
-		if len(addresses) > 0 && !slices.Contains(addresses, log.Address) {
-			return false
-		}
-		// If the to filtered topics is greater than the amount of topics in logs, skip.
-		if len(topics) > len(log.Topics) {
-			return false
-		}
-		for i, sub := range topics {
-			if len(sub) == 0 {
-				continue // empty rule set == wildcard
-			}
-			if !slices.Contains(sub, log.Topics[i]) {
-				return false
-			}
-		}
-		return true
-	}
-	var ret []*types.Log
-	for _, log := range logs {
-		if check(log) {
-			ret = append(ret, log)
-		}
-	}
-	return ret
 }
 
 func bloomFilter(bloom types.Bloom, addresses []common.Address, topics [][]common.Hash) bool {
