+++ melfilters/filter.go
@@ -1,0 +1,44 @@
+package melfilters
+
+import (
+	"math/big"
+	"slices"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+)
+
+// FilterLogs creates a slice of logs matching the given criteria.
+func FilterLogs(logs []*types.Log, fromBlock, toBlock *big.Int, addresses []common.Address, topics [][]common.Hash) []*types.Log {
+	var check = func(log *types.Log) bool {
+		if fromBlock != nil && fromBlock.Int64() >= 0 && fromBlock.Uint64() > log.BlockNumber {
+			return false
+		}
+		if toBlock != nil && toBlock.Int64() >= 0 && toBlock.Uint64() < log.BlockNumber {
+			return false
+		}
+		if len(addresses) > 0 && !slices.Contains(addresses, log.Address) {
+			return false
+		}
+		// If the to filtered topics is greater than the amount of topics in logs, skip.
+		if len(topics) > len(log.Topics) {
+			return false
+		}
+		for i, sub := range topics {
+			if len(sub) == 0 {
+				continue // empty rule set == wildcard
+			}
+			if !slices.Contains(sub, log.Topics[i]) {
+				return false
+			}
+		}
+		return true
+	}
+	var ret []*types.Log
+	for _, log := range logs {
+		if check(log) {
+			ret = append(ret, log)
+		}
+	}
+	return ret
+}
